+++ tests/core/test_lightning_optimizer.py
@@ -435,170 +435,3 @@ def test_lightning_optimizer_automatic_optimization_make_optimizer_step_2(tmpdir
 
         assert adam_zero_grad.call_count == 20
         assert sgd_zero_grad.call_count == 5
-
-
-@pytest.mark.skipif(not torch.cuda.is_available(), reason="test requires GPU machine")
-@pytest.mark.parametrize(["precision", "amp_backend"], [
-    pytest.param(16, "native"),
-    pytest.param(32, "native"),
-])
-@pytest.mark.parametrize('accumulate_grad_batches', [1])
-def test_parity_training_lightning_optimizer(tmpdir, amp_backend, precision, accumulate_grad_batches):
-    """
-    Test training with accumulated gradients with and within enable_pl_optimizer reaches the same weights
-    """
-    if accumulate_grad_batches > 1:
-        accumulate_grad_batches = np.random.randint(1, accumulate_grad_batches)
-
-    initial_weights = {}
-
-    class TestBoringModel(BoringModel):
-
-        def __init__(self, optimizer_name):
-            super().__init__()
-            self._optimizer_name = optimizer_name
-            self.losses = []
-            self.grads = []
-            self.on_before_zero_grad_count = 0
-
-        def on_before_zero_grad(self, optimizer):
-            self.on_before_zero_grad_count += 1
-            if self.layer.weight.grad is not None and not is_overridden("optimizer_step", self):
-                self.grads.append(self.layer.weight.grad.clone())
-
-        def configure_optimizers(self):
-            optimizer = getattr(torch.optim, self._optimizer_name)(self.layer.parameters(), lr=0.1)
-            return optimizer
-
-        def training_step(self, batch, batch_idx):
-            output = self.layer(batch)
-            loss = self.loss(batch, output)
-            self.losses.append(loss.detach().item())
-            return {"loss": loss}
-
-    TestBoringModel.training_epoch_end = None
-
-    max_epochs = np.random.randint(1, 3)
-    limit_train_batches = np.random.randint(11, 27)
-
-    expected_batches = max_epochs * limit_train_batches
-    limit_val_batches = 0
-
-    def train(enable_pl_optimizer, override=False, mock=False):
-        seed_everything(42)
-        if override:
-
-            def should_accumulate(trainer):
-                accumulation_done = (trainer.batch_idx + 1) == trainer.num_training_batches
-                is_final_batch = (trainer.batch_idx + 1) % accumulate_grad_batches == 0
-                return not (accumulation_done or is_final_batch)
-
-            # Note: global_step counts training_loop.optimizer_step
-            expected_global_step = expected_batches
-
-            class TestModel(TestBoringModel):
-                grad_checked = False
-
-                def optimizer_step(self, current_epoch, batch_nb, optimizer, optimizer_idx, closure,
-                                   on_tpu=False, using_native_amp=False, using_lbfgs=False):
-                    assert not isinstance(optimizer, LightningOptimizer)
-
-                    closure()
-                    if should_accumulate(self.trainer):
-                        return
-
-                    if using_native_amp:
-                        self.trainer.scaler.unscale_(optimizer)
-
-                    self.grad_checked = True
-                    assert torch.abs(self.layer.weight.grad).sum() > 0
-                    self.grads.append(self.layer.weight.grad.clone())
-                    if using_native_amp:
-                        self.trainer.scaler.step(optimizer)
-                        if mock:
-                            optimizer.step()
-                        self.trainer.scaler.update()
-                    else:
-                        optimizer.step()
-
-                    optimizer.zero_grad()
-                    if not mock:
-                        assert torch.abs(self.layer.weight.grad).sum() == 0
-
-            model = TestModel("SGD")
-            initial_weights["override"] = model.layer.weight.clone()
-        else:
-            expected_global_step = (expected_batches) // accumulate_grad_batches
-            if enable_pl_optimizer:
-                model = TestBoringModel("Adam" if mock else "SGD")
-                initial_weights["after"] = model.layer.weight.clone()
-            else:
-                model = TestBoringModel("AdamW" if mock else "SGD")
-                initial_weights["before"] = model.layer.weight.clone()
-
-        trainer = Trainer(
-            default_root_dir=tmpdir,
-            max_epochs=max_epochs,
-            limit_train_batches=limit_train_batches,
-            limit_val_batches=limit_val_batches,
-            enable_pl_optimizer=enable_pl_optimizer,
-            accumulate_grad_batches=accumulate_grad_batches if not override else 1,
-            amp_backend=amp_backend,
-            precision=precision,
-            gpus=1
-        )
-        trainer.fit(model)
-
-        assert np.abs(trainer.global_step - expected_global_step) <= 2
-        return model
-
-    before = train(False)
-    after = train(True)
-
-    assert torch.equal(initial_weights["before"], initial_weights["after"])
-    assert len(before.losses) == expected_batches
-
-    assert np.abs(len(before.grads) - (expected_batches // accumulate_grad_batches)) <= 2
-    assert not torch.FloatTensor(before.losses).isnan().any()
-
-    assert torch.equal(torch.FloatTensor(before.losses), torch.FloatTensor(after.losses))
-    assert before.on_before_zero_grad_count == after.on_before_zero_grad_count
-
-    for b_grad, a_grad in zip(before.grads, after.grads):
-        assert torch.equal(b_grad, a_grad), 'Grad parameters are different'
-
-    for b_w, a_w in zip(before.parameters(), after.parameters()):
-        assert torch.equal(b_w, a_w), 'Model parameters are different'
-
-    override = train(False, override=True)
-
-    assert torch.equal(initial_weights["before"], initial_weights["override"])
-    assert override.grad_checked
-    assert override.losses == before.losses
-    assert (override.on_before_zero_grad_count // accumulate_grad_batches) == before.on_before_zero_grad_count
-
-    for b_grad, o_grad in zip(before.grads, override.grads):
-        assert torch.equal(b_grad, o_grad), 'Grad parameters are different'
-
-    for b_w, o_w in zip(before.parameters(), override.parameters()):
-        assert torch.equal(b_w, o_w), 'Model parameters are different'
-
-    if precision == 32:
-
-        with patch("torch.optim.SGD.step") as mock_sdg_step, \
-             patch("torch.optim.Adam.step") as mock_adam_step, \
-             patch("torch.optim.AdamW.step") as mock_adamw_step, \
-             patch("torch.optim.SGD.zero_grad") as mock_sdg_zero_grad, \
-             patch("torch.optim.Adam.zero_grad") as mock_adam_zero_grad, \
-             patch("torch.optim.AdamW.zero_grad") as mock_adamw_zero_grad:
-
-            before = train(False, mock=True)
-            after = train(True, mock=True)
-            override = train(False, override=True, mock=True)
-
-        assert mock_sdg_step.call_count == (expected_batches // accumulate_grad_batches)
-        assert mock_sdg_zero_grad.call_count == (expected_batches // accumulate_grad_batches)
-        assert mock_sdg_step.call_count == mock_adam_step.call_count
-        assert mock_sdg_step.call_count == mock_adam_step.call_count
-        assert mock_sdg_zero_grad.call_count == mock_adam_zero_grad.call_count
-        assert mock_sdg_zero_grad.call_count == mock_adamw_zero_grad.call_count
